+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/XYUIMenuLiveSampleView.axaml.cs
@@ -1,0 +1,12 @@
+﻿using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views;
+
+public partial class XYUIMenuLiveSampleView : UserControl
+{
+    public XYUIMenuLiveSampleView()
+    {
+        InitializeComponent();
+        DataContext = new XYUIMenuSampleViewModel();
+    }
+}