+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI3/XYUI3-01-MenuBar/XYMenuBarItem.cs
@@ -1,23 +1,53 @@
 using Avalonia;
 using Avalonia.Controls;
+using Avalonia.Metadata;
 
 namespace XYUI.Avalonia.Controls;
 
 public sealed partial class XYMenuBarItem : Border
 {
     bool _building;
+    XYMenu? _menu;
+
     public static readonly StyledProperty<string> LabelProperty = AvaloniaProperty.Register<XYMenuBarItem, string>(nameof(Label), "");
     public static readonly StyledProperty<bool> IsActiveProperty = AvaloniaProperty.Register<XYMenuBarItem, bool>(nameof(IsActive));
     public static readonly StyledProperty<bool> IsHoveredProperty = AvaloniaProperty.Register<XYMenuBarItem, bool>(nameof(IsHovered));
+    public static readonly StyledProperty<bool> ShowChevronProperty = AvaloniaProperty.Register<XYMenuBarItem, bool>(nameof(ShowChevron));
+
     public string Label { get => GetValue(LabelProperty); set => SetValue(LabelProperty, value); }
+    public string Header { get => Label; set => Label = value; }
     public bool IsActive { get => GetValue(IsActiveProperty); set => SetValue(IsActiveProperty, value); }
     public bool IsHovered { get => GetValue(IsHoveredProperty); set => SetValue(IsHoveredProperty, value); }
-    public XYMenu? Menu { get; set; }
+    public bool ShowChevron { get => GetValue(ShowChevronProperty); set => SetValue(ShowChevronProperty, value); }
+
+    [Content]
+    public XYMenu? Menu
+    {
+        get => _menu;
+        set
+        {
+            if (_menu is not null) LogicalChildren.Remove(_menu);
+            _menu = value;
+            if (_menu is not null) LogicalChildren.Add(_menu);
+        }
+    }
+
     public event EventHandler? Activated;
+
     public XYMenuBarItem() { Classes.Add("xyui-menu-bar-item"); Build(); InitializeInteraction(); }
-    protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change) { base.OnPropertyChanged(change); if (!_building && change.Property != ChildProperty) Build(); }
+
+    protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
+    {
+        base.OnPropertyChanged(change);
+        if (!_building && change.Property != ChildProperty) Build();
+    }
+
     void Build()
     {
-        _building = true; Classes.Set("xyui-menu-active", IsActive); Classes.Set("xyui-menu-hover", IsHovered); Child = BuildVisual(); _building = false;
+        _building = true;
+        Classes.Set("xyui-menu-active", IsActive);
+        Classes.Set("xyui-menu-hover", IsHovered);
+        Child = BuildVisual();
+        _building = false;
     }
 }